Encoders: save stream labels/encoder configuration database.
A dedicated 'save' method is introduced so that stream labels and encoder configuration flags can be saved when SPL Engine/Streamer and/or NVDA terminates and to save flags when save keyboard command (Control+NVDA+C) is pressed. Unlike SPL Studio configuration database (ConfigHub), there is no caching involved (this can change), so the encoder version of save routine is simpler.