Encoders: do not write stream labels to disk after updating them.
When encoders support was part of SPL Utils global plugin, it was necessary to save stream label updates to disk. This was the case before widespread use of config save/reset routine from everywhere (so add-ons can take care of saving and loading/resetting settings). Now that a dedicated save method is used, there is no need to manually write stream label updates for encoders to disk - this will be taken care of through the new save method.