Encoder config: remove encoder config removed private flag.
The only time 'encoder config removed' private flag is set is when an encoder is removed from flags, and the flag itself was used to instruct NVDA to save stream labels database to disk if the only thing changed was flag sets. Now that the only time encoder settings database is written to disk is when it is actually saved, this flag is no longer necessary. This makes removal routines simpler.