Config UI: remove unused common alarm error dialog function. Re #125.
In the past, common alarm error dialog was used to announce an error message if a user tried to open another alarm dialog while an alarm dialog is active. Now that the scope has expanded to cover metadata streaming dialog, the old function is no longer used. Besides, Studio app module refers to a function that doesn't exist in splconfig module (alarm error was moved to config UI).