Broadcast profiles dialog: perform normal profile check before opening broadcast profiles dialog. Re #129.
When broadcast profiles was part of add-on settings screen, an explanatory text was shown if only normal profile was active or only in-memory config is used (and other profile functions removed if profiles are volatile). Now that broadcast profiles are managed from a separate dialog, normal profile and in-memory config check will be done prior to opening the dialog, as in these situations, broadcast profiles are meaningless. In short, this reverts an old commit from 17.10.