Broadcast profiles/close button: nullify '_selectedProfile' global flag for backwards compatibility. Re #129.
The global 'selected profiles' flag is used by profile-specific settings panels to determine which profile is selected. If thie flag is None, active profile is in use, otherwise the selected profile is different than what is being used. For backward compatibility, nullify this flag in close button handler for broadcast profiles dialog (this is temporary).