Profile specific panels (alarms/column announcements/metadata streaming): remove onDiscard method. Re #127.
The only real thing onDiscard method did was clear panel settings dictionary, which was needed when working with multiple profiles from add-on settings screen. With separation of broadcast profiles to its own dialog, there is no longer a need for it, thus remove onDiscard method. In case of column announcements panel, this also removes included columns set clearance routine, as it was used way back in 2018 when add-on settings screen was a pure settings dialog, not a multi-category settings (and now, the vestige of old dialog is finally removed).