Profile specific panels (alarms/column announcements/metadata streaming): refine onSave methods to remove traces of old profile selection handler routines. Re #127.
In the old days, when saving settings for alarms, column announcements, and metadata streaming panels, an extra check was performed to figure out which broadcast profile was active. This is no longer needed as the new settings screen will deal with a single profile, with broadcast profiles removed form add-on settings screen altogether.