Profile specific base panel: remove the panel altogether. Re #127.
In the old add-on releases, whenever broadcast profiles were changed from add-on settings, profile specific panels were notified through an action notification, and a base panel class was defined for this purpose. Now that broadcast profiles management is no longer the scope of add-on settings screen, the base panel and the actual notification handler is no longer needed. Thus remove the base panel altogether. Along with this change, the services used by the base panel wil be removed from individual profile specific panels.