Conf UI: introduce a dedicated function to open various settings panels. Re #125.
Move the functionality of config dialog open event to a new function that will be responsible for opening the settings panel requested, along with checking to make sure no other add-on settings dialogs are up. The config dialog event will then call this new function, passing in 'None' for panel argument (the new 'openConfigDialog' function takes a panel argument used to lok up the class to instantiate).
This is a foundation commit to bring the next commit on alarms panel to life.