Time-based profiles: remove next profile finder based on current time.
Next profile finder was used to return the next possible time-based profile based on current time - either queueing it or switching to the profile in question if NVDA was restarted in the midst of a broadcast. As time-based profiles are gone, there is no need to do this anymore, so remove next trigger profile finder.