Profile triggers: remove profile triggers and trigger timer variables.
Profile triggers were stored in a pickle-based dictionary, and a second copy was online for caching and comparison purposes (if both dictionaries are the same, do not write edits to disk). Along with this, trigger timer was used to actually switch to the chose nprofile and coordinate profile switches. As these are no longer used, remove them.