Config hub: change Window-Eyes option to NVDA (off) when looking up compatibility layer.
When compatibilty layer is set to 'wineyes', change it to 'off' (NVDA layout). This allows the command layout setting to be migrated from Window-Eyes to NVDA smoothly, and also helps NVDA see that options have changed, forcing config hub to write settings to disk. In the future this migration step will not occur.