SPL Engine and Streamer: do not bother cleaning up stream labels if it is None (no encoders initialized). Fixes #132.
Suppose a broadcaster opens Streamer or SPL encoders, then exits NvDA without focus on an encoder dntry. This will cause cleanup routine in encoders module to proceed as though stream labels is active when in fact it is None, causing TypeError (no assignment allowed to None type) to be raised. Thus do not proceed with cleanup action if stream labels is in fact None, meaning no more error sounds when restarting NVDA from encoders window when focused on something other than an encoder.