Encoders support: transplant encoder settings file error handling from Studio pap module to encoders module. Fixes #131.
Until now when encoder settings file becomes corrupt, a broadcast had to switch to Studio to fix this (and read the error message). This creates the following issues:
* Error when NVDA starts when focused on an encoder.
* Switches to an encoder window without switch to Studio first.
The old design was necessary as encoders support was part of SPL Utilities global plugin. Now that encoders support is part of SPL Engine and Stream app modules, it becomes easy to detect when encoder support should kick in. As a result, transplant the encoder file error handler to encoders module so it can be invoked when trying to load encoder settings for the first time. And if corruption is detected, start stream labels with an empty dictionary to prevent type error exception from being raised when trying to lookup settings.
This commit also serves as a test of Flake8-based linting hunt - F841 (variable defined but not used) was found in old encoder file handler routine, hence the transplant and to move away from old design.