Long-term support detection: use a temporary flag rather than a file when presenting LTS messages.
In 7.4 (2016), as part of preparing for LTS16 (15.x) release, a message was presented informing users to switch to lTS update channel. This was done by writing an empty text file during installation, which will then be picked up by Studio app module when it loads. Now that there is an easy way to remove keys form add-on settings, use temporary flag instead, similar to how time-based broadcast profiles feature deprecation message was shown (and removed).
This change also resolves Flake8 E265 (comment spacing) issue for lTS notification message, as this is now gone.