Add-on settings dialog: catch both RuntimeError and AttributeError.
In 18.09, with the advent of wxPython 4, wx.PyDeadObjectError was replaced by RuntimeError. To get around this, all exceptions were caught when destroying menu entry for SPL add-on settings. As wxPython 4 is here to say and based on test results, coupled with lint report fomr Flake8, catch both RuntimeError and AttributeError, as it is safe to do so now.
Although originally planned for 20.07 (Flake8 as this commit fixes E261 (inline spacing) and E262 (comments starting with #), prepare this for 20.06 as this change is long overdue.