Install tasks: import modules when running onInstall.
After speech refactor in a beta snapshot in April 2020, NVDA will emit a critical error message after installing the add-on. This was traced to both NVDA (imports) and the add-on install tasks (importing modules early). To get around this, import needed modules when running onInstall instead of at the top.