Encoders support: introduce a more general stream labels getter and setter.
As stream label getter and setter across encoder classes are really the same (the only difference being which stream labels map is used), implement a more general stream label getter and setter in base encoder class that will work with labels map as retured by subclasses. This reduces code duplication significantly from subclasses and paves the way for removing an unused getStreamLabel function defined at the module level, which has been inoperative for a while.
Note that getStreamLabel is not a property (although it can be viewed as such) to preserve backward compatibility and also to acknowledge the fact that a separate method is required to set stream labels.