SPL track item: catch COM error when trying to retrieve track name when a playlist is cleared.
Suppose a user clears the loaded playlist (Control+L) while focused on a track in Studio. If this happens, NVDA will play error tone (or appear to do nothing) due to COM error seen when trying to obtain track name (for a track that no longer exists). Therefore catch this error and return and empty string instead.