Place marker track checker: only list items can become place marker track.
In the past, place marker track would check any track passed in. The problem is that it looks up filename attribute, which ultimately uses column content getter which is defined for track list items, not other control types. Thus make sure that the 'track' being examined is indeed a list item before returning if it is indeed a place marker track. This fix is needed to resolve errors when trying to delete tracks in Studio's playlis viewer.