SAM encoders: set encoding flag to False when an error occurs while encoding is in progress.
Suppose an error occurs while encoding with SAM encoders. If this is detected while background monitoring is on, NVDA will not announce status updates and play the connection tone when the encoder is connected again. This is caused by 'encoding' flag not set properly when this scenario occurs. This is now reoslved - encoding flag will be set to False when this happens so connection tone and other updates can be announced.