Studio track item: remove left most column method. Re #142.
When Studio 5.0x was still supported, a dedicated left most column method was needed as tracks data structure was different between 5.0x and 5.1x. At that time (2014), it was assumed that Status column (0) can be fetched with just the MSAA name of the track item. Since then, column announcer was generalized, and a need to unify column navigation and announcement infrastructure was sought. Now that this is the case (Status column is column 0, thus allowing it to be treated just like any other column except for no Columns Explorer support), remove the ancient left most column method. This may come back if a future version of Studio changes how the first column is presented to users.