Vertical column navigation: move to appropriate column rather than announcing the correct column data. Re #142.
Now that child objects are used to represent columns, and since vertical navigation means moving to a navigator object, there is no need to call column announcer to announce correct column as determined by vertical column navigation setting (either navigate directly above or below or announce a specific column). Therefore vertical column navigation routine in report focus method has been edited to assign columns and use the built-in (and private) column navigation method to move to the 'real' column. This results in correct column highlighting if visual highlighting features are enabled in NVDA, but also results in screen jumping (something beyond the scope of the SPL add-on, a limitation to be documented in the future).