Studio track item/vertical column navigation: calculate column number based on saved column number variable from fake row behavior. Re #142.
As part of fake row behavior, a private variable holds which column is being looked at, which is consulted when moving vertically. Therefore use this variable to calculate column position rather than using a dedicated column number attribute. This is most prominent inside report focus routine as column position is needed to move to the correct cell.