Studio track item: remove custom column announcer method.
With the removal of status announcement flag, Studio track item version of column announcer became a 'tweaked' version of base column announcer. In reality, they do the same thing except for an extra call to indexOf method which is actually unnecessary. Therefore remove the derived version of column announcer (the only place this function will be invoked in Columns Explorer).