Encoders/connection reporter: exit connection reporter thread if both manual connect and background monitor flags are false.
Subtle optimization: in the past, as long as manual connect flag is on (manually connecting to a streaming server when a key is pressed) and background monitor is on (Control+F11), connection reporter thread would continue, and these were separate checks. This meant that background monitor check wasn't done until manual connect flag is turned off (after an encoder connects or an error occurs and the broadcaster asked NVDA to end connection reporting when this happens). Now both flags are checked at the same time (nor i.e. both manual connect and background monitor flags must be off) at the end.
This is applicable to both SAM and SPL/AltaCast encoder classes.