Config: do not save pilot features flag between resets and reloads. Re #144.
When reloading or resetting add-on settings, the only thing that was saved from reload and reset was pilot features flag. This made sense when this flag was actively used, as it was preferable to let users continue to experiment with early changes. However, as pilot features flag is being retired, along iwth the fact that this flag is not used anywhere, there is no need to guard this flag from reload and reset. Therefore remove this check.