Encoders: use a base implementation of connect follow-up actions.
When an encoder connects, connection reporter thread moves focus to Studio and/or plays the first selected track. Since both encoder classes come with identical follow-up action sequence, unify them under a base encoder class implementation. As the base implementation is a function, it will not be invoked (or rather, follow-up actions will not occur) if a track is playing and the broadcaster is located on Studio window.