Broadcast profiles dialog: rearrange dialog control placement.
Currently broadcast profiles dialog is laid out thus:
* Profiles list at the top
* Profile activate button below profiles list
* Buttons such as new, copy, etc.
* Close button
Although this allows profiles list to be placed on top, it doesn't align with NVDA Core and poses ergonomics issues, namely having to move down further to locate new button and so on. To align the experience with NVDA and to help with ergonomics (and clean up things a bit and make the dialog a bit amller), rearrange broadcast profiles dialog as follows:
* Horizontally divided into two parts
* Profiles list and activate button on the left (activate button is below profiles list)
* Buttons such as new anc copy on the right
* Close button at the bottom right corner below profile buttons.
Most of this change comes from NVDA Core (credit: Reef Turner from NV Access and other developers) and changes applied to suit the SPL add-on, namely addition of copy button and placing Triggers button below Delete button for simplicity.