SPL and AltaCast encoders: use unified encoder connect follow-up action from base class.
Same as SAM encoders: when SPL encoder family is connected, call the connection follow-up action routine defined in the base class.
Note: the private 'on connect' routine is queued in the main thread so SendMessage function can be carried out more effectively (not doing so in 2017 led to an issue where if a broadcaster asked Studio to play the first selected track when connected, it didn't do so).