Config: remove volatile config property.
To experiment with config saves, volatile config flag was introduced in 2017. With this flag on, once broadcast profiles were saved, they could not be saved. Since then, config save action allows emulating this setup by telling NVDA to not save settings on exit. Becasue of this, and with config.save method told to perform what used to be done as part of now removed volatile save method, volatile config flag (and the associated property) is no longer necessary.