Config hub: load settings from defaults dictionary if config in memory flag is active.
If config in memory flag is on, rather than calling copy profile function with SPL defaults dictionary as source profile, create a config obj with default settings included (via defaults.dict() method). This is cleaner as filename will not be defined at all, as the infile (first) argument to ConfigObj constructor can be a dictionary.