Config: get profile flags function is now part of Config Hub.
Meant to do it in 2017 timeframe: getProfileFlags function (returning profile flags such as 'active' for a given profile) is one of the last vestiges of 7.x style config pool mechanism. As this function is mainly used by broadcast profiles dialog and is a profile management routines, move this function inside ConfigHub structure.