Config: transfer pre-save function to ConfigHub.
The private pre-save function (_preSave) is one of the last vestige of old config pool methods used in 7.x days. As this function is private to the module, transfer the content (with appripriate edits) to ConfigHub module, along with editing variable names to fit what it is supposed to od (check if a profile should be saved/written to disk).