Revert SPL Assistant layer deduplication (January 22, 2017).
Suppose Studio app module is working yet SPL Utilities global plugin is not (perhaps a syntax error is encountered). Unless the error is with finally wrapper function, SPL Assistant will not operate s the global plugin class would not be online. Therefore restore finally wrap function in SPL app module to make the add-on components more independent, isolated, and resilient (global plugin is not really needed to use SPL add-on).