Startup dialogs: check minimal command-line switch inside startup dialogs function to allow displaying messages that should be displayed always.
For version deprecation and other important messages, startup dialogs function should be used, but this function will not be used if minimal command-line switch is in effect (no startup sound, no startup message, etc). Therefore move command-line switch check to inside startup dialogs function so that important messages can be announced (suppressing other startup messages if any) and to not present startup dialogs if minimal flag is on.