Column contents getter: use enumerate function rather than a pure for loop with index access. Re #148.
When fetching column contents as part of Playlist Transcripts, a for loop is used to go through column contents list with index access. This is most prominent when seeing if column has content. Rather than using index access, compare column content directly, and to modify the correct column position, use enumerate to step through column contents list (enumerate() function iterates through a iterable with an optional start position, yielding the index and the item in question).