Cart Explorer: use enumerate function to locate both cart entry and position. Re #148.
Until now when populating carts dictionary, after making sure the cart entry is indeed defined, list.index method was used to obtain entry poosition. This position variable had a '1' added to it because the first cart entry is F1, not F0. To avoid this situation and to work directly with both cart entry and position, use enumerate method with start parameter set to 1.