Config hub init: remove extra values from broadcast profiles, too.
At the moment extra values (mostly removed settings) are checked when loading normal profile. But there might be times when a profile-specific setting is deprecated and removed. To handle such a case, extend extra values check to broadcast profiles as well. Just like normal profile, if deprecated keys are found, cache the broadcast profile in question and remove deprecated keys.
This will be backported to LTS20 for code consistency.