SPL Utils: check if Studio window is even visible before moving system focus to it. Fixes #150.
Reported by a user: when trying to move to Studio while it is minimized, NVDA will announce 'unavailable' in the end. This is caused by the fact that when Studio window is minimized, the window itself is invisible. Therefore check if the Studio window is visible before attempting to switch to it (in a way, the old 'Studio is minimized' message has been resurrected with an additional detail that the window has been minimized to system tray).