Studio playlist viewer item: emulate building SysListView32.ListItem name pieces. Re #149.
If column header inclusion setting is off, build name pieces (column contents separated by semicolon). This is applicable to both regular report focus routine as well as when custom column inclusion/order combo is in effect. Note that regular name string builder (at the top of the report focus routine) is a temporary work for 20.11 and will be removed in a future commit.