BuildVars: remove minor version flag from compatibility flag.
Reported by a member of NVDA Add-ons list: an add-on cannot be installed if minimum/last teted version is in the form 'year.major.minor' but it works if 'minor' flag is not specified. Therefore remove the minor version flag i.e. last tested version = 2020.4.0 -> 2020.4.
This will be backported to 20.09.x.