SPL confspec: add a dedicated module to hold SPL add-on config specifications. Re #151.
A fix for Flake8 E501 for confspec and to follow NVDA Core's lead: add a dedicated config spec module to record SPL add-on settings specs, similar to NVDA's config spec module for hosting NVDA settings specs. The ultimate purpose of this module is to tell Flake8 to ignore this module so E501 (line too long) will not be shown, as some add-on settings (particularly column announcement list) can become a long string.