ui.message (Flake8 E501): split ternary operation due to line length.
Ternary operations are employed in parts of the add-on. While elegant, it leads to long lines. This is the case for announcing Studio status and library scan progress announcement. For these (defined in Studio app module), convert ternary operations to traditional conditional (if/else). This is most noticeable for status announcement as status variable will be split if message verbosity is set to 'advanced'.