Playlist viewer item/vertical column announce: locate the correct column if vertical column announce value is an actual column header and column display order has changed.
When moving vertically, make sure to take into account the fact that display order may change. Previously only the header index was fetched, meaning that the column to be reviewed is not really the same one as what the display says e.g. vertical column value is 'Artist' yet column position for this column is different. Note that column order array is not really a Python list, so convert it into one and then fetch index if possible (no need to catch value error as long as playlist viewer shows 18 columns).
This will be backported to LTS20.