Library scan announcer: queue the announcement function to main thread to avoid errors with braille handler thread.
Because library scan reporter runs inside a dedicated thread, library scan announcer must be queued to main thread, more so if a braille display is in use. This resolves a situation where NVDA would stop announcing library scan updates if the setting is set to announce scan progress and a braille display is in use.