Studio playlsit viewer item: add a command to toggle screen column order announcement (NVDA+V). Re #153.
Add a dedicated keyboard command to toggle screen column order announcement setting (column announcement settings panel), assigning it to NVDA+V. NVDA+V is used to toggle screen layout in browse mode documents, and it is used in this add-on to toggle column presentation for tracks - similar in idea. For consideratior of braille users, update braille message after toggling screen column order setting.
This will be backported to lTS20 for consistency.