Studio app module: reset cart edit timestamp list to None.
Inconsistent resource deallocation: when Studio app module exits, cart edit timestamps are set as zero's (a list is retained). This is even more so when cart explorer was not even invoked. Whereas cart explorer does reset timestamps to None but app module terminate method doesn't, edit the terminate method to reset timestamps to nothing i.e. set the list to None for consistency.