Studio/track comments: catch KeyError exception when trying to remove track comments. Re #156.
Consider the following: a user presses Alt+NVDA+C three times to open track comment dialog for a track with no comment defined. If OK is clicked without typing anything, KeyError will be raised because the filename for this track doesn't exist. Therefore catch this exception and move on.
This will be backported to LTS20.