Playlist duration/snapshots: include an empty string to a tuple of segue checks used for calculating total duration.
In the past, only None or 00:00 were checked when deciding to add segue in total duration. These days an empty string ('') is retuend if segue cannot be obtained. Therefore include an empty string to a tuple of values to exclude when calculating total duration.
This will be backported to LTS20.