Playlist snapshots: convert segue string into an integer for easier minimum and maximum duration comparisons.
Python's string comparison looks at value of the character (such as ASCII value). For example, a segue of '1:00:00' is less than '30:00' because '1' is less than '3' (accorindg to ASCII). This causes wrong tracks to be flagged as shortest and longest duration (segue). Therefore convert segue into an integer to ease min/max comparison (minimum and maximum will also be converted into integers and later converted into a time format string before presenting shortest and longest tracks and their duration).
NOte that a different variable will be used to hold integer segue because segue is also used to calculate total duration of a playlist later.
This will be backported to LTS20.