ConfigHub/create profile/new profile dialog: profile name argument renamed, name should not be None.
Caught by Mypy: new profiles is a set of strings, and it breaks if None is passed in when creating a profile. Therefore do not take None as a parameter. Also, for consistency with rename and delete operations, profile name argument is now claled 'name'.
This will be backported to LTS20.