ConfigHub/reset: convert confirmation flag to boolean.
Confirmation flag is initially a bool. However, when an instant switch profile is active, assignment may not work properly becasue it wil be a boolean between a bool and an integer (switch flags is an int vector but is treated as an int). Therefore convert confirmation flag to a bool just to make sure and to satisfy static checkers.
This will be backported to LTS20.