Studio app module/place marker: place marker flag will now store track filename, not filename index.
In the past, place marker flag stored filename index and track filename so the former can be cached. However, it is prone to a critical error: place marker flag can be reassigned to anything, and place marker routines do not check validity of filename index value. Besides, indexOf function is called in places to obtain filename, and unless it is moved, a track's filename alone can serve as a handy place marker flag. Therefore place marker will be track's filename alone.
This will be backported to lTS20 for consistency.