Type annotations: use __future__.annotations flag for backward and forward compatibility. Re #155.
Annotate functions and variables to make it clear as to what is what and what type it is. As a first step, use __future__.annotations import to guarantee forward and backward compatibility (to be removed in Python 3.10 version). This is better than importing uppercase versions of containers from typing module as these are deprecated in Python 3.9.
The overall type annotations are guided by Mypy, a static type checker.