ConfigHub: re-initialize ChainMap.maps as a list of ConfigObj. Re #155.
Mypy: re-initialize ChainMap.maps as a list of ConfigObj. This complies with ChainMap specification in that maps is a list of mappings, and ConfigObj, as a dictionary/section, is indeed a mapping. To acknowledge that ConfigHub is powered by ChainMap, call super method, and then re-initialize self.maps. This resolves attribute definition errors from Mypy when accessing ConfigObj attributes.