Track finder: pass in an empty columns list if no column is specified.
Now that track finder/track locator requires a list of columns (integers), pass in an empty string if columns are not specified (NvDA will then look for artist and title). Also, in track finder method inside Studio app module, columns list will not be None - just check columns, as it will then check length of columns list i.e. if search columns are specified.
This will be backported to LTS20.