ConfigHub/delete profile: obtain index for the profile to be deleted after swapping it with normal profile.
Mypy: contrary to its name, swap profiles method can return the index to the former profile if requested. However this raises errors with Mypy which insist on correct type for this method. It turns out this issue stems from delete profile method where it wants to know the index for the profile name if it is the current active broadcast profile. To resolve this, delete profile method will no longer ask for profile index. Instead, it will swap profiles if deleting the active broadcast profile, and then it will ask for profile index.