Config: do not proceed with closing config and switching between profiles if SPLConfig is gone.
Attribute access/Mypy: guard agaisnt SPLConfig becoming None, otherwise NVDA will attempt to close a nonexistent config and switch to broadcast profiles it doesn't know about. When SPLConfig does become None, raise a RuntimeError but will not be caught by code that calls affected functions to communicate the severity of this issue.
This will be backported to lTS20.