Config/attribute access: return from initializing config if SPLConfig turns out to be None. Re #155.
After openConfig is called, SPLConfig would be something but Mypy assumes config will be None. Therefore to satisfy Mypy and to guard against this scenario, return from initializing ad-on config if SPLConfig does become None. At least this allows Studio app module to load, it may result in errors, which can be a clue to developers that soemthing odd must have happened with ConfigHub.