Studio/track starts: return early if track starts variable from Studio API returns none. Re #155.
Mypy: return early if track starts value from Studio returns None. Also, make sure to annotate track starts variable when first used because it can be either an integer or None, and is used a second time (if it is an integer) as part of a divmod/time announcer operation.