Studio API: return None if Studio dies im the midst of a freeze when studio exits.
NVDA will appear to freeze when Studio exits, especially more so while Studio API monitoring or library scan monitoring is in progress. If this happens, Studio API will return 0 because the Studio window handle is gone. Therefore check for window handle once more before returning Studio API result, changing the result to None if Studio is indeed gone (handle is NULL). This resolves a host of bugs related to Studio API monitoring, particularly NVDA saying '0 items in the library' when in fact Studio is gone.
This will be backported to LTS20.