SPL track item/columns explorer: simplify columns explorer by fetching column header and content from child objects instead of using SysListView32 API.
Because SysListView32 objects are internally seen as a table by NVDA, individual columns are child objects of list items. Therefore instead of fetching column header and content through SysListView32 API, fetch the child object where the specific column header is defined. This is most prominent if the track item class defines 'explore columns' list (Studio playlist viewer item, Track Tool, Creator) where column position can change depending on index of the header text (the columns explorer list can raise ValueError if the header is not found, and to avoid this, membership is checked). This significantly simplifies columns explorer script as it is now a child fetcher instead of dealing with SysListView32 raw column header and content. Also, the nonexistent header (typically seen if a header is not defined in older SPL suite) is checked only if custom columns explorer property is defined. Note that SysListView32 API will be employed in LTS20 to guarantee stability.