ConfigHub: ignore profile cache when saving profiles.
Do not consult profile cache when writing profile data to disk as profiels are not even cached at all. Note that this breaks profile save function in that profiles are not saved to disk unless a reset takes place (to be resolved by writing profiles to disk regardless of profile flags, to be committed after removing cache mechanism).