ConfigHub: remove new profiles set.
Although not apparent: suppose a profile rename takes place, and somehow the new name comes from a set of new profiles for this session. This is not an obvious bug, but it highlights use of something that is not useful anymore (another purpose of new profiles set was to save them to disk). Due to this subtle bug, coupled with removal of cache mechanism means new profiles set is no longer useful, therefore it is gone.