Studio 6.0: use display text as status bar item labels.
Studio 6 changes the screen layout. More importantly, it does not include accessibility labels for status bar items as it used to. Among other things, this causes NVDA to not announce status changes such as automation. Thankfully, display text can serve as a label, so tell NVDA to treat this as status bar item label at the cost of not announcing them in the beginning.