Studio 6.0: handle remainig/song ramp window changes.
One of the places that changed layout is remaining/song ramp text. As opposed to being next to each other, element label and the actual timer is separate (opposite ends of the window objects). Thankfully there is a 'Track Starts' label next to remaining/song ramp text, so use it to detect Studio 6 layout, otherwise name change event for remainig/song ramp is same as Studio 5.x.