SPLConfig: globalVars.appArgsExtra -> globalVars.unknownAppArgs in NVDA 2022.1.
With the possibility of letting add-ons (including SPL add-on) specify extra command-line arguments, extra app arguments are stored in unknown app args list in NVDA 2022.1. To support old and new releases, use a dedicated list to hold either the old (appArgsExtra) or new (unknownAppArgs) command-line switch list to which the add-on will look for SPL add-on specific switches (mostly restricting to using normal broadcast profile and/or loading profile from memory).