SPL utils (security): disable the global plugin altogether in secure mode.
Discussion on NVDA add-ons list: globalVars.appArgs.secure is not strong enough - even if the global plugin constructor returns early if secure mode is on, commands such as switching to Studio window still works. Therefore use a solution provided on add-ons list to disable the global plugin altogether (using a decoator that returns the the empty global plugin or the actal SPL Utils global plugin if secure mode is on or off, respectively). While this issue also applies to app modules, it doesn't affect all parts, but for sake of completeness, the decorator method will be applied to app modules in the next set of commits.