Playlist transcripts: only 'view transcripts' will be available in secure mode.
Althougg copyin transcripts to clipboard and saving data to a file are disabled in secure mode, the actual options in playlist transcripts dialog (SPL Assistant, Shift+F8) are not. Therefore only 'view transcripts' option will be available in this dialog. This is done by restructuring transcript actions so the first option (view transcripts) is always added to the actions list and other two options (copy to clipboard, save to file) are added if not in secure mode.