SPL Utils: register Streamer as an alias of SPL Engine.
NVDA 2022.2 introduces a convenience function in ap module handler to register an executable as an alias of an app module, allowing contents of one app module to be imported from another (thereby removing 'from appMod import *' workaround). In SPL add-on, this applies to SPL Streamer as it is an alias of SPL Engine as it uses components from the latter. This also means buffer size label workaround will be removed.