SPL Studio/SPL Assistnat/speak on demand: rename selffinish to self.script_fnish to allow speak on demand to take effect wihle SPL Assistant is active.
To allow SPL Assistant commands (mostly information with a few action commands) to be spoken, define speak on demand attribute in the 'finish' command. This is because while perfomring SPL Assistant commands, the next script will be the 'finish' function. As part of this work, rename self.finish method to self.script_finish without a gesture argument (at least to fool NVDA that this script is' going to do some work' when in fact it was marked as a script to get speak on demand mode to work in layer commands).