SPL Controller: global plugin version of SPL window handle varialbe is no longer used.
As splbase module includes its own attribute to record Studio window handle, there is no need for the global plugin to carry this information. With this commit, the SPL utilities global plugin now dependson Studio ap module/splbase module for full functionality, an architecture that was implied but never became a thing for more than ten years (this design was documented in add-on internals with the statement that the global plugin is subordinate to the Studio app module).